requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
oauthlib==3.2.2
openai==1.54.4
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==1.3.1
schedule==1.2.2
sniffio==1.3.1
tqdm==4.67.0
tweepy==4.14.0
typing_extensions==4.12.2
urllib3==2.2.3
